requirements.txt
# Campaign Finance Analysis App - Python Dependencies
# Updated: 2025-12-10

# Core Framework
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3

# Visualization
plotly==5.18.0
altair==5.2.0

# Data Processing
openpyxl==3.1.2
pyarrow==14.0.2              # For Parquet support

# Configuration & Environment
pyyaml==6.0.1
python-dotenv==1.0.0

# Web Scraping & Data Download
requests==2.31.0
beautifulsoup4==4.12.3
urllib3==2.1.0

# AI/LLM Integration (for chat feature)
openai==1.10.0
anthropic==0.18.0

# Statistical Analysis
scipy==1.11.4

# Network Analysis (for donor-committee graphs)
networkx==3.2.1

# Geospatial (optional - for maps)
folium==0.15.1
streamlit-folium==0.16.0

# Progress Bars
tqdm==4.66.1

# Date/Time handling
python-dateutil==2.8.2

# Logging
colorlog==6.8.2
